Split of the doctor codebase
The code of `qmk doctor` started causing `flake8` complexity warnings.
This commit moves the helper functions into a common module
`qmk.os_helprs` and the Linux-specifics one into `qmk.os_helpers.linux`.
This also makes adding future, OS-specific checks easier.
